Activity 6 - Relative Jumps
Task description:
Change the interpreter so that it uses relative jumps.

Many virtual machines and CPUs use relative jumps: an instruction
jump X does not mean "jump to position X", but "jump X positions ahead"
(A backward jump uses a negative X).

Among other advantages, relative jumps make the code relocatable:
that means that the code for a function works correctly regardless of
where it is loaded in the final code.