Activity 8 - 'elseif'
Task description:
Add elseif to the language.
The general syntax for an if becomes like this:

  if c1 then s1
  elseif c2 then s2
  ...
  elseif c_n then s_n
  else s_n+1

Note that the AST doesn't need to know about elseif. A statement like:

  if c1 then s1
  elseif c2 then s2
  else s3
can generate the same tree as the statement:

  if c1 then s1
  else { if c2 then s2 else s3 }
That means that you have to change only the parser to implement this construction.