Activity 11 - Default Argument
Add default arguments to the language. Functions that accept a default
argument must be declared like here:

  function foo (a, b, c = exp) { ... }

Only the last argument for a function can have a default value.
The extra = exp after the parameter c signals that foo can be called
with two or three arguments.
If called with only two arguments, exp must be evaluated to provide
the default value for the c parameter. When foo is called with three
arguments, exp is not evaluated.